Remover simulação de digitação e usar colagem direta
