+++ manifest.json
@@ -13,9 +13,6 @@
   "content_scripts": [
     {
       "js": [
-        "simulation/base.js",
-        "simulation/tool.js",
-        "simulation/typing.js",
         "content.js"
       ],
       "matches": [
@@ -36,16 +33,7 @@
     {
       "resources": [
         "icon-128.png",
-        "icon-34.png",
-        "simulation/base.js",
-        "simulation/tool.js",
-        "simulation/typing.js",
-        "simulation/comum/simulador.js",
-        "simulation/hibrida/simulador.js",
-        "simulation/instantaneo/simulador.js",
-        "simulation/robusta/simulador.js",
-        "simulation/teste-simulacao.html",
-        "simulation/teste-simulacao.js"
+        "icon-34.png"
       ],
       "matches": ["<all_urls>"]
     }
