+++ popup.html
@@ -226,28 +226,7 @@
       </select>
     </fieldset>
 
-    <!-- Teclado Simulado -->
-    <h2>Simulação de Teclado</h2>
-    <div id="typingOptions">
-      <p style="font-size:0.75rem; margin-bottom:0.5em;">
-        Modos simulados reproduzem atrasos e pequenas falhas para maior realismo.
-      </p>
-      <label for="typingMode">Modo de digitação</label>
-      <select id="typingMode">
-        <option value="instant">Instantâneo</option>
-        <option value="simulate">Comum</option>
-        <option value="hybrid">Híbrida</option>
-        <option value="robust">Robusta</option>
-        <option value="imitator">Imitador</option>
-        <option value="rewrite">Apagar/Reescrever</option>
-        <option value="rewriteRandom">Reescrever Aleatório</option>
-        <option value="templateFast">Template Rápido</option>
-        <option value="templateSlow">Template Lento</option>
-      </select>
-      <button id="openTestPage">Abrir Página de Teste</button>
-    </div>
-
-    <!-- API AdsPower -->
+      <!-- API AdsPower -->
     <fieldset>
       <legend>Integração</legend>
       <label for="adsPowerApi">URL da API AdsPower</label>
