+++ test/run-tests.js
@@ -1,3 +1,2 @@
 require('./test-selector.js');
 require('./utils.test.js');
-require('./simulateKey.test.js');
